Basis for MTA analysis.  Finds the last saturday date and uses that to find the MTA data.  Fixes a column error and creates a histogram of the hour of the day of updates.
Created a method that parses the # Rides for each station in the PATH trains.

Created a method that calculates the change in entries/exits for a station in the PATH train network.

Optimized the saturdayDate method.

Placed pd.read_csv into its own method.  Surrounded it in try/except block.

Grouped pathTrains by station and date.  Created graphs with plotly